pbs-316--feat: 답변 페이지 구현
